requirements.txt
fastapi
uvicorn
httpx
sqlalchemy
psycopg2-binary
pydantic
google-adk
google-genai
python-dotenv